requirements.txt
docker>=4.0.0,<7.0.0
dbus-python>=1.2.0
